testd/requirements.txt
selenium==4.10.0
webdriver-manager==4.0.0
playwright==1.36.0
requests==2.31.0
imap-tools==1.0.0
python-dotenv==1.0.0
2captcha-python==1.2.0
loguru==0.7.0

# OCR dependencies for captcha solving
pytesseract==0.3.10
opencv-python==4.8.0
numpy==1.24.3
Pillow==10.0.0